backend/requirements.txt
Backend is live. Let me fix requirements.txt and build the React frontend.
Action: file_editor create /app/backend/requirements.txt --file-text "fastapi==0.115.0
uvicorn==0.30.6
httpx>=0.28.1
python-dotenv==1.0.1
"
Observation: Overwrite successful: /app/backend/requirements.txt

Installing dependencies from requirements.txt